Gilberto Agostinho:
I did quite a lot of work on textures and materials lately and I'd like
to ask if someone could please take a look on it and maybe commit it to
FGDATA if judged positively.

Starting with commit 0da71d2288836df05548f5f343fe3977ebb2955f of my
repository (see
https://github.com/gilbertohasnofb/new-regional-textures/commits/master
) I did:

- improved urban effect for the Latin American city texture
- improved the Mediterranean city texture (and also better urban effect)
- created new textures for asphalt and concrete runways
- created a new industrial texture (as well as winter version and urban
effect)
- created winter versions for European 18th century city and town textures
- applied urban effect to European 18th century city and town textures
- created a new material in which small areas around 80 large cities
around the world in which the street lights have higher density (1000
for mega cities such as NY and Tokyo, 2000 for large cities such as
Munich and Dubai, while keeping the default 10000 for all the rest of
the globe)

I also created a texture with skid marks for Thorsten's new effect, but
we are still working on this one and this may not even be used.
Everything else is ready to go.

I know this is a lot of things to test, and as Thorsten pointed out in
the forum I it would be better if next time I submit my work in several
chunks (which makes testing easier). So I could break this new work like
that if necessary, but here is a reasonable plan for anyone willing to
test it as it is now:

- start at LIRN at noon during summer. Check the new asphalt runway.
Take off fly around the city centre (it's just in front of the airport).
Test the improved European 18th century city and town textures, as well
as the new industrial texture. Change the environment to winter and
check these three textures once again.
- start at KLAX at noon. Check the new concrete runway. Take off, change
the time to midnight and check the street lights (Los Angeles is a "mega
city" so should be very bright).
- start at SBRJ at noon and check the improved Latin American city
texture. Take off, change the time to midnight and check the street
lights (Rio de Janeiro is a "large city" so should be quite bright).
- start at DTTG @ runway 06 and check the town by the left of the
airport for the improved Mediterranean city texture.

Also, there are plenty of screenshots of this work starting with this
post here: http://forum.flightgear.org/viewtopic.php?f=5&t=26031&start=150#p261282