Route manager: Support list of waypoint specifications in dialog input
Allow multiple, space-separated waypoint specifications to be added to a
route, rather than having to type each one and press the add button
each time.

If no waypoint is selected as the insertion point, show a message
and leave the waypoint input intact, rather than just clearing it.

Update the help label to reflect the change and include a note of how to
create offset waypoints using a radial and distance.