Multiplayer Dialog fixes.
Change to use the ports defined in the property tree when connecting from the dialog.
Ensure properties for ports are set to a valid (non zero, not null) value on dialog initialisation.
Add ability to change receive port from dialog firstly to see what it is and secondly to allow multiple fg sessions