Translations: remove unwanted leading and trailing whitespace
It's not that leading or trailing whitespace in translated strings is
forbidden *in principle*, but I think it is unwanted in most of these
cases. Remove these spaces that could cause confusion, possibly useless
work for translators and/or ugly displays.

The only cases where the spaces seem to have been intentional are the
trailing ones in menu labels (Translations/*/menu.xml), in order to
obtain right alignment of keyboard shortcuts. However, there is only one
case in English where the spacing looks approximately correct
(File -> Reset). In all other cases, it looks awful. And this poor way
of obtaining right alignment gives bad results in other languages, where
translators use the same number of spaces as in English.

Bottom line: if right alignment of keyboard shortcuts is indeed desired
in the menu labels (I think it is), it should be done in a smarter
way...