Translations: rename Translations/en to Translations/default
This is mainly to make it clear that the "translation" in (after this
commit) Translations/default, a priori corresponding to en_US, has a
special status: it acts as the source describing the set of translatable
strings, each with its id (an XML tag name) and numeric index (a
PropertyList index, often 0 but not always).

In the upcoming i18n infrastructure, the files in Translations/default
will be used to:
  - detect all translatable strings (~ string extraction);
  - detect obsolete translated strings (those in XLIFF files but not in
    Translations/default anymore);

This will allow a script to update the XLIFF files containing real
translations, i.e., update the source strings and possibly remove
obsolete entries.

So, the files in Translations/default have a special "master" status in
this process, and are technically quite different from those in
Translations/de, Translations/es, Translations/fr, etc. The latter XML
files (those not in Translations/default) will be replaced with their
XLIFF equivalent, by the way (conversion script already working).

See discussion at
<https://sourceforge.net/p/flightgear/mailman/message/35910579/>.