Update all non-French XLIFF translation files
The reason French is treated specially is that I completed the
translation up to 100% coverage in Qt Linguist as a test case for the
new translation infrastructure, and have no desire to convert this
translation work to the legacy FlightGear-specific XML format. But of
course, this complete translation is in XLIFF 1.2 format (since it is
saved by Qt Linguist) and thus can't be used by FlightGear yet.

On the other hand, for the other languages, no translator ever modified
the XLIFF files, therefore the best way to proceed for is that people
continue to work with the legacy FlightGear-specific XML format until
the new translation infrastructure is ready (this way, their work is
visible in FlightGear). The Python-based tools in FGMeta allow one to
automatically convert from this legacy format to the XLIFF 1.2 format,
as explained in fgmeta/python3-flightgear/README-l10n.txt. Which is
precisely what this commit does.