LOD dialogs rework.
Big rework of the Adjust LOD ranges dialog to hopefully make it easier to understand, and harder to setup so that models aren't visible by accident.

All range entry is now via sliders.

The sliders are deltas from the previous value.

There is now the option to specifically choose either low detail or high detail AI/MP models. *NOTE* that for models without a fallback index the low detail model will be the blue yellow glider.

There is now clear indication of what happens for the Meters or Pixel modes.

Added a defaults button that will restore a reasonable set of default values.

The properties that were used have been renamed to ensure that sensible defaults will be picked, ideally we'd have a way to migrate things like this but I don't think we do, so this is a reasonable alternative.