Revert "Nasal/screen.nas: move log window to top-left and left-align the text."
It's always been at the top; and although it is clear that the intent of this change was to improve things it highlights the need to make more substantial improvements to the way on screen messages are handled.

What we probably should have is an area that has a transparent dark background, a fixed size, supports scrolling, can be left open (or re-opened) and moved persitently around the screen.

This reverts commit 82f3eb45b9265d893b7617f9ccef7da99123d410.