Fix Nasal failures on clean install.
Previously not all Nasal modules were defined in defaults.xml.  This
meant that the first time FlightGear was run it was possible for
Aircraft Nasal files to reference a module that had not been loaded yet.

This commit simply defines all the Nasal modules ahead of time and loads
them, so all modules are available for Aircraft Nasal files
immediately.