Fix multiple openings of FG1000 windows
Previously on the first call to display an FG1000 PFD/MFD from
the Debug menu worked, because only one index was used, which was
subsequently deleted when the window was closed.

Now a new PFD/MFD instance is created each time the Debug menu is used.
This isn't ideal, and probably exposes Nasal memory issues if used
many times, but is fine for the moment.