Nasal library load order
Control the order in which fgdata/Nasal/*.nas files are loaded
on startup.

Previously all the files were loaded in order based on filename.
This was problematic for aaa.nas which might want to call
props.getNode(), and generally resolved by using
_setlistener("/sim/signals/nasal-dir-initialized")

This change allows us to configure a list of Nasal files that are
loaded first on startup - loadpriority.xml.