I18N: update the French FlightGear-nonQt.xlf
I only started to translate the new 'weather-scenarios' part, however
I'm too ignorant about the proper terms and phrases used in the field.
Thus, it would be much better if someone more competent in aviation and
weather terminology could review and finish the 'weather-scenarios'
part. The other parts are now fully translated (not the Qt launcher, but
that should be done in FlightGear-Qt.xlf).