Merge Emesary bug fixes 2020.x
Transfer encoding rewritten to handle negative numbers properly and fix scaling.

Second set of changes from Nikolai

Emesary MP bridge: The fixes have in common that they prevent a
notification from being processed multiple times.

OutgoingBridge: When transmitting queue, make sure transmitted
messages cannot remain in the queue when not completely emptied, and
thus be sent again, and eventually fill up the queue.

Also allow smaller messages to sent even though there was not room for
a larger message (the removal of the 'break' command).

To prevent coding the same message again with the 'break' removed,
store an already coded string that was rejected due to too little room
inside the message so when there is room for it, it was coded only
once.

OutgoingBridge: Stop recalculating MessageExpiryTime when there is no
room for a notification. If it has asked to expire after expire time,
let it expire.

IncomingBridge: Make sure listeners on emesary[x] properties are
removed when a mp aircraft is invalid, otherwise a hidden bridge will
sit in background and process same notifications a new bridge is
already processing.

IncomingBridge: Change order of setting IncomingMessageIndex to ease
debugging in reciever.

IncomingBridge: Since a bridge class can handle multiple emesary
properties on same aircraft, use a vector per aircraft to store the
bridge instances in, to make sure Remove can be called on all
instances listening to same aircraft.

EncodeFixedDouble: Clamp what is sent on max byte value instead of
scale.