Make ufo invisible over MP again
Before fallback models were introduced, the ufo
was invisible to other MP pilots that were not in a
ufo themselves.

When fallback models were introduced, at long range
the ufo was replaced by the glider model, and so became
visible.

This change makes it invisible again by using the ufo as its
own fallback model.  This is safe because the ufo model is
very lightweight and also shipped with fgdata.