Checklist: <and> multiple entries in <condition>
Previously a <condition> block was assumed to contain
a single child, and other children were ignored, largely due to the PUI code.

Most developers would expect multiple children to be evaluated
as if surrounded by an <and> block, which is how the tutorial code
behaves.

This change wraps the children in an <and> making it consistent.