Colin Geniet:
There are 3 relevant render bins for aircraft models:
- 0: default
- 10: depth sorted, automatically selected for transparent objects.
- 111: depth sorted, explicitly set by some transparent effects (hud, glass)

The model-default effect does not set <render-bin> on purpose, to allow
automatic selection of bin 0 or 10. So my second patch which added <render-bin>
to model-default is wrong.
It is only model-transparent which should explicitly set bin 111
to interact well with hud/glass effects.