WS30: Add vpb directory to suffixes for terrasync
This allows WS3.0 to sync scenery over terrasync.

At present there is no UI to enable this, as WS3.0
is not available on the main terrasync servers.  So
it needs to be manually enabled by setting

--prop:/sim/rendering/scenery-path-suffix[98]/enabled=true

Note that WS3.0 has a dependency on the airport
data that is stored under the Terrain suffix.

Thanks to merspieler for setting up a test
terrasync server to test this.