Fix placements not being removed when destroying a canvas
Canvas placements are being added / removed through a listener on "placement" node adds / removals on the canvas's node.
Previously, when the canvas's del method was called, the canvas's node was removed, but not its "placement" subnodes.
This caused the canvas to be destroyed without the placements being removed.
This in turn causes problems when setting up a canvas with the same placement(s).
This is fixed by this commit by removing the "placement" nodes separately before destroying the canvas.