HDR: Add use-clustered-uniforms to forward pass
Transparency shaders in the HDR forward pass use clustered uniforms for
lighting, but the necessary uniform samplers aren't set. On Mesa they
default to 0 which results in GL_INVALID_OPERATION from glDrawElements
due to the same texture unit being used by multiple texture targets (3D
and 2D), and so transparent objects don't render.

Now that all pass builders support clustered uniforms, we can add a
use-clustered-uniforms property node to the forward pass to specify the
texture units to bind to. The same texture units are used as for the
opaque shading pass, except pointlights which uses 16 instead of 13,
since 13 is in use for something else.