Fix autopush bug that causes nasal errors in MapStructure,
Fix overwrite of the global id() function with an integer, which will break anything that attempts to use it later, and MapStructure is the most common victim of this, since it relies heavily on id() to check for the presence of map elements.

Changing it into a local variable (with var) prevents this, keeping this id scoped locally.

error manifests as

[ALRT]:nasal      Nasal runtime error: Fatal error in Nasal call: function/method call on uncallable object

Fix originated by tdammers.