HDR: Add compute shader for histogram calculation
Add a compute shader for histogram calculation as a much faster
alternative to the histogram-column and histogram-aggregate fragment
shaders when OpenGL 4.3 is available.

Two new passes are added for fast histogram generation. The
histogram-clear pass uses a trivial fragment shader to reset the
histogram texture ready for the histogram-compute pass of the new
"compute" type to perform atomic adds on the bins in the histogram
texture for each pixel in the histogram-bins texture.

The switching between slow fragment shader and fast compute shader is
done at the effect technique level, with all 4 effects (2 slow, 2 fast)
having predicates added to their techniques to enable compute shaders
only when both supported (OpenGL >= 4.3) and enabled via a property
which defaults to true (/sim/rendering/hdr/histogram/compute). This
allows the compute shader to be dynamically disabled with the internal
property browser for testing.

In a particularly bad case for the fragment shader approach at high VR
resolutions this change reduces the 8.7ms GPU time required for
histogram calculation using fragment shaders (histogram-column and
histogram-aggregate, excluding histogram-bin which has now been split
out of histogram-column) down to around 50us with the new compute shader
(histogram-clear and histogram-compute).

The compute shader itself uses a number of techniques to perform
efficiently, and is currently tuned for an AMD Radeon RX 6700XT. It
operates in workgroups of 256x2 invocations, each workgroup calculating
8 partial histograms in parallel covering 256x32 pixels (16 vertical
pixels binned per invocation).

The partial histograms exist in per-workgroup shared memory to reduce
inter-workgroup contention of atomic adds. There are multiple partial
histograms to reduce contention between invocations within the
workgroup. They are initialised before use, summed after use, and
atomically added to the main histogram in parallel, reducing the number
of atomic operations to the main histogram bins, thereby reducing
contention.