canvas map: Fix crash from lakebed runways
Nasal error in canvas map caused FG to shutdown when selecting an
airport with lakebed runways (e.g. KEDW) in the airport dialog.
This is due to the surface type colour being undefined.

This is a regression introduced by 9538bfff9 (Fix uses of surface type,
2024-03-15), which removed the fallback entry SURFACECOLORS[0].

Restore the fallback value, and add entries for unusual runway types:
12 = lakebed, 13 = water, 14 = snow, 15 = transparent
(cf. https://developer.x-plane.com/article/airport-data-apt-dat-12-00-file-format-specification/#codes)

Clearly mark 0 as being a fallback entry. This was not the case
previously, which is likely why 9538bfff9 removed it (it is an invalid
code in the X-plane format).

Reported by Nikolai V. Chr.:
https://sourceforge.net/p/flightgear/codetickets/2893/