HDR/vert shaders: Transform frame buffer texcoords
Use a new mvr_raw_texcoord_transform_fb() function, defined in mvr.vert
and mvr.frag, to transform frame buffer texture coordinates from view
space to screen/texture space. This is needed to allow for correct
sampling of intermediate frame buffer textures using a side-by-side
layout, e.g. SceneView stereo VR mode, and soon to be supported
single-pass multiview rendering modes.

The implementation uses the OSGXR_VERT_MVR_TEXCOORD() &
OSGXR_FRAG_MVR_TEXCOORD() macros when defined by osgXR.

A new QUAD_TEXCOORD_RAW definition can be passed to quad.vert and
quad_matrix.vert to provide a new raw_texcoord varying which is
untransformed.

The raw_texcoord is needed for atmos_aerial_perspective.frag, which
manipulates the raw texture coordinates to determine which depth slice
is being drawn before performing the transform.

The raw_texcoord is also needed for postprocess.frag, which uses it to
calculate the redout distortion (with the viewport transform taking
place per fragment after distortion), the redout vignette, and the
ev100 debug (so the bar at the bottom is duplicated on each view).

The raw_texcoord is also needed for debug-histogram since the histogram
being sampled is not multiview but the visualisation is still drawn to
both views.

Finally raw_texcoord is also needed for debug-aerial-perspective, since
the preview is divided into 8x8 cells. Sampling of the multiview
aerial-perspective buffer is handled in the next commit since it is a
fixed-size-per-view buffer rather than scaling to the compositor size.