HDR: Clear envmap to fix undefined value issues
If the aircraft starts just below the terrain, for example the UFO,
parts of the hdr envmap won't be initialised, and may contain negative
values. When these values are used they may cause whole blocks of pixels
to go black (Linux, Mesa 24, AMD GPU).

Therefore clear the envmap before drawing into it to ensure undrawn
values are safe to use.