HDR: Add aerial-perspective visualisation
Add a visualisation of the aerial-perspective buffer, enabled with the
property /sim/rendering/hdr/debug/show-aerial-perspective.

This draws each of the 32 depth slices of the buffer in a 8x8 grid, with
the in-scattering colours and transmittance values shown on alternating
rows. Red single-pixel-wide lines are used to separate each
cell.

This is useful for debugging, especially with VR where the buffer may
need to be treated slightly differently to framebuffers due to it being
per-view and also fixed resolution.