HDR: Do not clear the envmap every frame
Related to commit c07b6327

Clearing the envmap this way was problematic because <render-condition>
only modifies the osg::Camera's cull mask, which does not prevent the
clearing operation from happening every frame.

The skydome is now capped at the bottom, so the envmap should be
correctly initialized at startup even when the camera is below the
terrain. See SimGear commit 1bd5dad3.