HDR/ws30.frag: Transform water normals to view space
For WS3 with HDR, tangent space water normals were transformed by a
fixed rotation matrix in generateWaterNormal() then treated as being in
view space. This makes the normals rotate with the view, and so the
lighting rotates with it too.

Drop the fixed transform in generateWaterNormal() so that tangent space
normals are returned, and construct a TBN (tangent-binormal-normal)
matrix using the input normal and the fg_CameraViewUp uniform. This is
used to transform the water normals from tangent space into view space,
which appears to make them behave correctly.