HDR/redout: Simplify redout_distort()
The distortion calculation in redout_distort() can be simplified without
changing the result. By multiplying the uv vector itself rather than the
calculated length we avoid the sqrt() to calculate the length, as well
as the vec2(sin(), cos()) to calculate the normalised vector, which also
makes the atan() redundant.