HDR: Use tessellation shaders for WS30
The tessellation shaders procedurally subdivide the terrain geometry
according to the distance to the camera. They expect a GL_PATCHES
primitive of 16 vertices. The 4 central ones form the actual terrain
quad, while the 12 neighbouring vertices are used for bicubic
interpolation.

The procedural detail is shaped by noise functions of different
wavelengths. Features that are too small to displace the actual geometry
is later used for normal mapping in the fragment shader.

The fragment shader has been temporarily simplified to better debug the
tessellation shaders.