Continuous UV range across scenery
This change generated a second set of texture coordinates that
are continous over the whole earth surface.

For each 1x1 degree block they range from (1,1) at each corner to (0,0)
in the center.

E.g.

        (1,1)          (0,1)             (1,1)
        +--------------+--------------+
        |                             |
        |                             |
        |            (0,0)            |
    (1,0) |              +              | (1,0)
        |                             |
        |                             |
        |                             |
        +--------------+--------------+
    (1,1)          (0,1)             (1,1)

Due to the way that the 1x1 blocks tile, they end up being continuous.

They are intended to be used as input into noise systems that require
a continuous set of UV coordinates.