HDR: Make aerial-perspective buffer multiview
Change the type of the aerial-perspective buffer to 3d-mvr so that it
can fit a 64x64x32 buffer for each view.

To generate the contents, the atmos-aerial-perspective-compute pass has
a new "multipass" multiview mode set to dispatch the compute shader once
per view. The atmos-aerial-perspective-layered is not updated yet, but
VR users should be able to use the compute version.

The compute shader is slightly updated to convert the raw texture
coordinate for the aerial perspective slice (derived from
gl_GlobalInvocationID) to frame buffer coordinates that can be used for
calculating the view space vector. It also uses a new helper
ap_get_voxel_view_offset() to offset the coordinates and write to the
correct view of the buffer.

For now mvr_raw_texcoord_transform_fb() behaves the same as
mvr_raw_texcoord_transform_buf(), which is not actually the same thing,
but as long as the views are the same size it'll do for now.