HDR: Fix cloud environment map rendering
Clouds aren't being drawn to environment maps for me on AMD hardware
with mesa. The Effects/cloud's hdr-envmap shaders fail to link because
aerial_perspect_envmap.glsl refers to compute_inscattering() in
get_aerial_perspective(), but atmos.glsl and its dependent shaders
aren't attached to the fragment stage.

I would have expected it to work since get_aerial_perspective() isn't
used from the fragment stage, but adding it back fixes the clouds in the
environment maps.

It is assumed that Effects/cloud-static.eff is also affected so it is
similarly fixed, though the advanced rendering settings dialog is a
little difficult to read so I haven't been able to work out how to
trigger that effect.