l10n: fully update the English FlightGear-nonQt.xlf
The first target text in each translation unit has been set to the
source text. Everything has then been marked as approved in Linguist.
The effect of this is that when strings with plural forms appear in the
default translation and migrate to this file, they will stand out and
will be easy to translate.