Rename atc.xml to atc.no_translate.xml in Translations/default
This aligns with FGData commit 979624790ea in 'next'. In order to allow
aircraft and add-ons to use their own translation categories (aka
contexts, or resources), the Python i18n scripts in FGMeta have been
recently changed to act on all XML files in Translations/default, except
those ending with '.no_translate.xml'. As we don't want strings from
atc.xml to propagate to the .xlf files, we need this renaming.