VR: Enable 360 mouse mode
Connect up a 360 mouse mode, unlinking the 3D mouse cursor from the OS
mouse cursor (which is constrained and hidden), allowing it to be moved
around freely in the scene, even out of the current view, without the OS
mouse cursor accidentally leaving the window, which is useful for VR
where the user can't even see the OS window.

This mode can be toggled explicitly when VR is enabled (or if the VR
mouse cursor is always shown) with Ctrl+Tab, or with Shift+Right-Click.
If user presence is supported by the OpenXR runtime, 360 mouse mode can
also be enabled/disabled automatically when the VR headset is put on or
removed. This can be controlled with a new option in the VR options
dialog: "Enable 360 mouse when VR in use".