XMLDialog.nas: enable localized <format> strings with plural forms
If the <format> config node of a widget has translate="y", it should (as
usual) contain a translation ID and the <?FlightGear-tr ...?> processing
instruction that gives the “engineering English” version of the <format>
string. If this is the case and the widget “local value” is an integer,
then this value is used to choose the correct plural form in case the
processing instruction declares the translatable (format) string with
has-plural="true".

This goes with the following FlightGear commit:

  PUICompatObject: allow translations whose plural status is determined at runtime