Add 3D mouse cursor models
Add mouse cursor models in Blender and AC3D formats, along with an
effect and XML files to apply the effect. The Blender files are included
as all the meshes are generated using modifiers from simple curves,
and in the case of the hand models an armature and skinning to allow
posing of the hand.

The models are referenced in VR/cursors.xml which is included in the
main property tree at /sim/vr/config/cursors/.