Add 3D mouse cursor direction cue model
Add mouse cursor direction cue model in AC3D format, along with an XML
file to apply the mouse cursor effect. No Blender file is included as
the mesh is trivially simple.

The model is referenced in VR/cursors.xml so that it can be picked up by
flightgear when the upcoming 360 mouse mode is in use, giving the user a
visual cue as to the direction the mouse cursor is even if out of sight.